Optimize `skip(n: Long)` in BufferedInputStream
The DWARF code now calls many times the skip method
in `BufferedInputStream`.
It had an implementation note of being inefficient.
It was because was reading one byte at a time instead
of skipping multiple bytes at once.
After optimizing it (basically porting the read(array) implementation
without the actual copy to a target array), the benchmark
went from ~46ms to ~36ms