JUnit: populate `sbt.testing.Event.throwable` and `sbt.testing.Event.duration`.
If a Throwable associated with the test event is available, bubble it up through the SBT's test interface using the slot dedicated to just such information: `sbt.testing.Event.throwable`.

If duration associated with the test event is available, bubble it up through the SBT's test interface using the slot dedicated to just such information: sbt.testing.Event.duration;
if duration is not available, populate the slot with `0`.

It seems a shame to throw away information that is literally in our hands; also, this change:
- simplifies integration with runners that assume that test failure is always accompanied by a `Throwable` (e.g., Gradle) and
- avoids weird-looking negative duration values in test reports.