GenericProcess: put exit code in CompletableFuture
It then serves both as a cached value and as a basis for onExit() calls.
For the latter, we no longer register multiple CompletableFuture objects
in GenericProcessWatcher but only the process itself.

Also, replace checkResult with getExitCodeImpl and in UnixProcessGenX,
synchronize not the entire waitFor call but check the cached exit code
first and then call osWaitForImpl which is synchronized.