[javalib] GenericProcess: drain/close each stream separately (#4540)
* UnixProcessGen2: don't copy stderr on redirect

When stderr is redirected to stdout, we need ensure stderr is the same
file descriptor as stdout (that was done correctly), and after that the
exposed error stream is empty (as everything is available from stdout).
However, this last part wasn't done, so let's rectify that.

* GenericProcess: drain/close each stream separately

To avoid potential deadlocks, such as when the child process is writing
to stderr while we attempt to drain stdout, let's process each stream
separately and asynchronously.