[javalib] UnixProcessGen2: don't create unused pipes, close all unused descriptors (#4550)
* UnixProcessGen2: conditionally create all pipes

They are only needed in case when redirect is Redirect.PIPE.
Otherwise, the descriptors are leaking.

Previously, we were only somewhat conditionally creating stderr.

* UnixProcess: close child pipe ends in the ctor

* UnixProcessGen2: child closes all fds except 0,1,2

Despite the comment in the code, dup2 doesn't close the original file
descriptor.

And if the child doesn't properly close some write file descriptors, it
might be messing with the kernel pipe logic detecting EOF. (Reportedly,
the kernel will only report EOF on a pipe if there are no open write
descriptors and no data.)