[javalib] GenericProcess: no synchronized, ignore reaped (#4588)
* UnixProcessGen2: refactor osWaitForImpl

Turn it into a val function reference, change underlying linux and bsd
methods to return success, and merge the logic into the old infamous
askZombiesForTheirExitStatus.

* GenericProcess: no synchronized, ignore reaped

One of the reasons we used to synchronize waitFor and setExitCode was
the race condition between learning the child exited, then reading its
exit status and, finally, setting it in the completion object.

So, make the following changes:
- tolerate inability to reap a child which was presumably signaled, as
  it may have been reaped by another;
- do not "invent" an exit code if a child cannot be reaped; this way,
  only the thread which succeeded in reading the exit code will set it.