[javalib] BsdOsSpecific: fix bug in kevent struct ident/data (#4601)
The `ident` and `data` fields are not pointers, they are integral types
large enough to contain a pointer to `uint`/`int`, respectively. In SN
parlance, they are USize and Size.

Continuing to define them as Ptr[_] types makes them impossible to use.
A call to `kevent` might store an integer in that field, and not a Ptr
which led to the hack of using kevent64_s to manipulate them instead.
And kevent64_s aligns with kevent only on 64-bit platforms.

Similarly, the final field of kevent64_s is not a pointer but an array
with 2 elements, so let's represent it as two separate primitive fields.

Also, add field accessors for kevent and kevent64_s.