[javalib] GenericProcessWatcher: move most to nested object (#4643)
* GenericProcessWatcher: modify removeSomeProcesses

Instead of key-value entries, use a function on values only.

* GenericProcessWatcher: move most to nested object

Move processes field, watcher thread and some methods to a nested object
so that we both encapsulate this functionality and also delay thread
initialization without using `lazy`.

This will allow us to add another "lazy" field to the Processes object
later on.