WeakReference: fix JDK compliance w/ clear/enqueue
The usual, JDK behaviour of `clear()` and `enqueue()` methods is a bit
different than what was implemented in scala-native. Let's fix that and
make sure to test for consistency between JDK and SN.

For that, move WeakReferenceTest to shared, but also require JDK9.
JDK9 contains Cleaner and thus will allow us to test advanced logic.